requirements.txt
Flask==2.0.1
tensorflow==2.15.0
Pillow==10.3.0
numpy==1.23.0
werkzeug==2.0.3